chore(core): move redisclient, pubsub, lock to infra
