+++ modules/core/infra/distributedlock/redis_lock.go
@@ -1,30 +1,12 @@
-package lock
+package distributedlock
 
 import (
 	"context"
-	"errors"
 	"time"
 
 	"github.com/redis/go-redis/v9"
 	"github.com/sky-as-code/nikki-erp/modules/core/logging"
 )
-
-var (
-	ErrAcquireLockFailed = errors.New("acquire lock failed")
-)
-
-type DistributedLock interface {
-	Acquire(ctx context.Context, key string, ttl time.Duration) (bool, error)
-	AcquireWithRetry(
-		ctx context.Context,
-		key string,
-		ttl time.Duration,
-		retryCount int,
-		retryDelay time.Duration,
-	) (bool, error)
-
-	Release(ctx context.Context, key string) error
-}
 
 type RedisDistributedLock struct {
 	logger      logging.LoggerService
