+++ modules/core/infra/pubsub/interface.go
@@ -1,0 +1,11 @@
+package pubsub
+
+import "context"
+
+type Publisher interface {
+	Publish(ctx context.Context, channel string, message any) error
+}
+
+type Subcriber interface {
+	Subscribe(ctx context.Context, channel string) (<-chan []byte, error)
+}
