+++ modules/core/infra/pubsub/redis_pub_sub.go
@@ -12,21 +12,23 @@ type RedisPubSub struct {
 	redisClient *redis.Client
 }
 
-func NewRedisPubSub(logger logging.LoggerService, redisClient *redis.Client) *RedisPubSub {
-	return &RedisPubSub{
+func NewRedisPubSub(logger logging.LoggerService, redisClient *redis.Client) (Publisher, Subcriber) {
+	r := &RedisPubSub{
 		logger:      logger,
 		redisClient: redisClient,
 	}
+
+	return r, r
 }
 
-func (this *RedisPubSub) Publish(ctx context.Context, topic string, message any) error {
-	this.logger.Debug("Publishing message to topic", logging.Attr{"topic": topic, "message": message})
-	return this.redisClient.Publish(ctx, topic, message).Err()
+func (this *RedisPubSub) Publish(ctx context.Context, channel string, message any) error {
+	this.logger.Debug("Publishing message to topic", logging.Attr{"topic": channel, "message": message})
+	return this.redisClient.Publish(ctx, channel, message).Err()
 }
 
-func (this *RedisPubSub) Subscribe(ctx context.Context, topic string) (<-chan []byte, error) {
-	pubSub := this.redisClient.Subscribe(ctx, topic)
-	this.logger.Debug("Subscribing to topic", logging.Attr{"topic": topic})
+func (this *RedisPubSub) Subscribe(ctx context.Context, channel string) (<-chan []byte, error) {
+	pubSub := this.redisClient.Subscribe(ctx, channel)
+	this.logger.Debug("Subscribing to topic", logging.Attr{"topic": channel})
 	_, err := pubSub.Receive(ctx)
 	if err != nil {
 		this.logger.Error("Failed to subscribe to topic", err)
@@ -48,4 +50,3 @@ func (this *RedisPubSub) Subscribe(ctx context.Context, topic string) (<-chan []
 func (this *RedisPubSub) Close() error {
 	return this.redisClient.Close()
 }
-
