--- modules/core/lock/index.go
@@ -1,9 +1,0 @@
-package lock
-
-import (
-	deps "github.com/sky-as-code/nikki-erp/common/deps_inject"
-)
-
-func InitSubModule() error {
-	return deps.Register(NewRedisDistributedLock)
-}
