--- modules/core/redisclient/index.go
@@ -1,11 +1,0 @@
-package redisclient
-
-import (
-	deps "github.com/sky-as-code/nikki-erp/common/deps_inject"
-)
-
-func InitSubModule() error {
-	return deps.Register(
-		NewRedisClient,
-	)
-}
